+++ tests/torch2/function_hook/quantization/test_fq_lora.py
@@ -37,8 +37,8 @@ from nncf.torch.quantization.layers import AsymmetricQuantizer as AQ
 from nncf.torch.quantization.layers import LoraMixin
 from nncf.torch.quantization.layers import SymmetricQuantizer as SQ
 from tests.cross_fw.shared.paths import TEST_ROOT
-from tests.torch.ptq.test_weights_compression import ShortTransformer
 from tests.torch.test_models.synthetic import LinearModel
+from tests.torch.test_models.synthetic import ShortTransformer
 from tests.torch2.utils import compare_with_reference_file
 from tests.torch2.utils import to_comparable_nx_graph
