+++ tests/torch2/function_hook/quantization/test_quantizer_config.py
@@ -22,9 +22,9 @@ from tests.cross_fw.test_templates.models import NNCFGraphToTestDepthwiseConv
 from tests.cross_fw.test_templates.models import NNCFGraphToTestSumAggregation
 from tests.cross_fw.test_templates.models import NNCFGraphTransformer
 from tests.cross_fw.test_templates.test_quantizer_config import TemplateTestQuantizerConfig
-from tests.torch.ptq.helpers import get_depthwise_conv_nncf_graph
-from tests.torch.ptq.helpers import get_single_conv_nncf_graph
-from tests.torch.ptq.helpers import get_sum_aggregation_nncf_graph
+from tests.torch2.function_hook.quantization.helper import get_depthwise_conv_nncf_graph
+from tests.torch2.function_hook.quantization.helper import get_single_conv_nncf_graph
+from tests.torch2.function_hook.quantization.helper import get_sum_aggregation_nncf_graph
 
 
 class TestQuantizerConfig(TemplateTestQuantizerConfig):
