+++ tests/torch2/function_hook/quantization/test_weights_compression.py
@@ -15,6 +15,8 @@ import pytest
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
+from transformers import AutoModelForCausalLM
+from transformers import AutoTokenizer
 
 import nncf
 from nncf import BackupMode
@@ -38,8 +40,6 @@ from nncf.torch.quantization.quantize_functions import pack_uint4
 from nncf.torch.quantization.quantize_functions import unpack_int4
 from nncf.torch.quantization.quantize_functions import unpack_uint4
 from tests.cross_fw.test_templates.template_test_weights_compression import TemplateWeightCompression
-from tests.torch.ptq.test_weights_compression import AWQActLinearModel
-from tests.torch.ptq.test_weights_compression import AWQLinearModel
 from tests.torch.test_models.synthetic import ShortTransformer
 from tests.torch.test_tensor import cast_to
 
@@ -91,6 +91,85 @@ class LinearModel(torch.nn.Module):
 
     def forward(self, input):
         return self.linear(input)
+
+
+class AWQActLinearModel(nn.Module):
+    def __init__(self, with_multiply=False, n_layers=8):
+        super().__init__()
+        self.with_multiply = with_multiply
+        self.n_layers = n_layers
+
+        def create_linear_layer():
+            weight_tensor = torch.arange(0, 64, dtype=torch.float32).reshape(8, 8) - 32.0
+            linear_layer = nn.Linear(8, 8, bias=False)
+            linear_layer.weight = nn.Parameter(weight_tensor)
+            return linear_layer
+
+        self.linear_emb = create_linear_layer()
+
+        self.linears_1 = nn.ModuleList()
+        if self.with_multiply:
+            self.linears_2 = nn.ModuleList()
+
+        for _ in range(n_layers):
+            self.linears_1.append(create_linear_layer())
+            if self.with_multiply:
+                self.linears_2.append(create_linear_layer())
+        self.linear_lm_head = create_linear_layer()
+
+    def forward(self, x):
+        out = self.linear_emb(x)
+
+        for i in range(self.n_layers):
+            node1 = F.relu(self.linears_1[i](out))
+            if self.with_multiply:
+                node2 = torch.selu(self.linears_2[i](out))
+                out = node1 * node2
+            else:
+                out = node1
+
+        out = self.linear_lm_head(out)
+        return out
+
+
+class AWQLinearModel(nn.Module):
+    def __init__(self, is_int8=False):
+        super().__init__()
+        self.is_int8 = is_int8
+
+        self.linear1 = self.get_linear_layer(0.01 * torch.arange(0, 64).reshape(8, 8) + 0.05, is_int8)
+        self.linear2 = self.get_linear_layer(0.01 * torch.arange(0, 64).reshape(8, 8) + 0.05, is_int8)
+        self.linear3 = self.get_linear_layer(0.01 * torch.arange(0, 64).reshape(8, 8) + 0.05, is_int8)
+        self.linear4 = self.get_linear_layer(0.01 * torch.arange(0, 64).reshape(8, 8) + 0.05, is_int8)
+        self.linear5 = self.get_linear_layer(0.01 * torch.arange(0, 64).reshape(8, 8) + 0.05, is_int8)
+        self.linear6 = self.get_linear_layer(0.01 * torch.arange(0, 64).reshape(8, 8) + 0.05, is_int8)
+
+    def get_linear_layer(self, weights_data, is_int8):
+        if not is_int8:
+            linear_layer = nn.Linear(weights_data.shape[1], weights_data.shape[0], bias=False)
+            linear_layer.weight = nn.Parameter(torch.tensor(weights_data, dtype=torch.float32))
+        else:
+            qw = torch.tensor(weights_data, dtype=torch.uint8).float()
+            zp = torch.tensor([2**7], dtype=torch.uint8).float()
+            scale = torch.ones((weights_data.shape[0], 1), dtype=torch.float32)
+            weights = (qw - zp) * scale
+            linear_layer = nn.Linear(weights_data.shape[1], weights_data.shape[0], bias=False)
+            linear_layer.weight = nn.Parameter(weights)
+
+        return linear_layer
+
+    def forward(self, x):
+        node1 = self.linear1(x)
+        node2 = self.linear2(x)
+        node_multiply = node1 * node2
+
+        node3 = self.linear3(node_multiply)
+        node4 = self.linear4(node3)
+        node5 = self.linear5(node3)
+        node_multiply_2 = node4 * node5
+
+        node6 = self.linear6(node_multiply_2)
+        return node6
 
 
 class FunctionalModel(torch.nn.Module):
@@ -475,6 +554,18 @@ class TestPTTemplateWeightCompression(TemplateWeightCompression):
             )
         }
 
-    def test_error_message_for_invalid_group_size(self):
-        # TODO (dokuchaev) fix wrapping.
-        pytest.xfail("Known issue with wrapping")
+
+@pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float16])
+def test_half_precision_models(dtype):
+    model_id = "hf-internal-testing/tiny-random-OPTForCausalLM"
+    model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=dtype)
+    tokenizer = AutoTokenizer.from_pretrained(model_id)
+    inputs = tokenizer("dummy_input", return_tensors="pt")
+    compress_weights(
+        model,
+        group_size=2,
+        mode=CompressWeightsMode.INT4_SYM,
+        scale_estimation=True,
+        awq=True,
+        dataset=nncf.Dataset([dict(inputs)]),
+    )
